Revert "Converted to use PEAR's HTTP_Request2, since, according to http://pear.php.net/package/HTTP_Request/ HTTP_Request has been superceded by Request2. "
This reverts commit e18f3c2419f2f033e235744273f16982e4d10661.

There's no debian package for HTTP_Request2 and it's difficult to
install the package without updating a bunch of other stuff.  I'm
making the executive decision that it's too new to switch over to this
version of the library; if I'm having these problems others will as
well.  Tell me if you think I'm wrong and we can roll forward again.